test-param-wait-system: fix race between tasks
Calling RequestName with gdbus-tool to signal the tested binary that a
name is owned is racy: since the name is only owned very shortly, the
watch might not have been established before the name is already gone
again.

Fix this by holding on to the name a bit longer with the test-own-name
helper (add a --system flag to that).

A better fix would be to only start the second task when the watch has
been established. dbus-test-runner always starts tasks in parallel
though and changing this would further complicate its command line
interface.